classifier/requirements.txt
fastapi[standard]
uvicorn
onnxruntime
numpy
pillow
